+++ pubs/commands/add_cmd.py
@@ -12,6 +12,7 @@ from .. import apis
 from .. import pretty
 from .. import utils
 from .. import endecoder
+from ..command_utils import add_doc_copy_arguments
 from ..completion import CommaSeparatedTagsCompletion
 
 
@@ -26,37 +27,39 @@ def parser(subparsers, conf):
     parser = subparsers.add_parser('add', help='add a paper to the repository')
     parser.add_argument('bibfile', nargs='?', default=None,
                         help='bibtex file')
-    parser.add_argument('-D', '--doi', help='doi number to retrieve the bibtex entry, if it is not provided', default=None, action=ValidateDOI)
-    parser.add_argument('-I', '--isbn', help='isbn number to retrieve the bibtex entry, if it is not provided', default=None)
+    id_arg = parser.add_mutually_exclusive_group()
+    id_arg.add_argument('-D', '--doi', help='doi number to retrieve the bibtex entry, if it is not provided', default=None, action=ValidateDOI)
+    id_arg.add_argument('-I', '--isbn', help='isbn number to retrieve the bibtex entry, if it is not provided', default=None)
+    id_arg.add_argument('-X', '--arxiv', help='arXiv ID to retrieve the bibtex entry, if it is not provided', default=None)
     parser.add_argument('-d', '--docfile', help='pdf or ps file', default=None)
     parser.add_argument('-t', '--tags', help='tags associated to the paper, separated by commas',
                         default=None
                         ).completer = CommaSeparatedTagsCompletion(conf)
     parser.add_argument('-k', '--citekey', help='citekey associated with the paper;\nif not provided, one will be generated automatically.',
                         default=None, type=p3.u_maybe)
-    parser.add_argument('-L', '--link', action='store_false', dest='copy', default=True,
-                        help="don't copy document files, just create a link.")
-    parser.add_argument('-M', '--move', action='store_true', dest='move', default=False,
-                        help="move document instead of of copying (ignored if --link).")
+    add_doc_copy_arguments(parser)
     return parser
 
 
-def bibentry_from_editor(conf, ui, rp):
+def bibentry_from_editor(conf, ui):
+
     again = True
-    bibstr = templates.add_bib
+    bibentry_raw = templates.add_bib
+    decoder = endecoder.EnDecoder()
+
     while again:
         try:
-            bibstr = ui.editor_input(initial=bibstr, suffix='.bib')
-            if bibstr == templates.add_bib:
+            bibentry_raw = ui.editor_input(initial=bibentry_raw, suffix='.bib')
+            if bibentry_raw == templates.add_bib:
                 again = ui.input_yn(
                     question='Bibfile not edited. Edit again ?',
                     default='y')
                 if not again:
                     ui.exit(0)
             else:
-                bibentry = rp.databroker.verify(bibstr)
+                bibentry = decoder.decode_bibdata(bibentry_raw)
                 bibstruct.verify_bibdata(bibentry)
-                # REFACTOR Generate citykey
+                # REFACTOR Generate citekey
                 again = False
 
         except endecoder.EnDecoder.BibDecodingError:
@@ -82,30 +85,29 @@ def command(conf, args):
     citekey = args.citekey
 
     rp = repo.Repository(conf)
+    decoder = endecoder.EnDecoder()
 
     # get bibtex entry
     if bibfile is None:
-        if args.doi is None and args.isbn is None:
-            bibentry = bibentry_from_editor(conf, ui, rp)
+        if args.doi is None and args.isbn is None and args.arxiv is None:
+            bibentry = bibentry_from_editor(conf, ui)
         else:
-            if args.doi is not None:
-                bibentry_raw = apis.doi2bibtex(args.doi)
-                bibentry = rp.databroker.verify(bibentry_raw)
-                if bibentry is None:
-                    ui.error('invalid doi {} or unable to retrieve bibfile from it.'.format(args.doi))
-                    if args.isbn is None:
-                        ui.exit(1)
-            if args.isbn is not None:
-                bibentry_raw = apis.isbn2bibtex(args.isbn)
-                bibentry = rp.databroker.verify(bibentry_raw)
-                if bibentry is None:
-                    ui.error('invalid isbn {} or unable to retrieve bibfile from it.'.format(args.isbn))
-                    ui.exit(1)
-                # TODO distinguish between cases, offer to open the error page in a webbrowser.
-                # TODO offer to confirm/change citekey
+            bibentry = None
+            try:
+                if args.doi is not None:
+                    bibentry = apis.get_bibentry_from_api(args.doi, 'doi', ui=ui)
+                elif args.isbn is not None:
+                    bibentry = apis.get_bibentry_from_api(args.isbn, 'isbn', ui=ui)
+                    # TODO distinguish between cases, offer to open the error page in a webbrowser.
+                    # TODO offer to confirm/change citekey
+                elif args.arxiv is not None:
+                    bibentry = apis.get_bibentry_from_api(args.arxiv, 'arxiv', ui=ui)
+            except apis.ReferenceNotFoundException as e:
+                ui.error(e.message)
+                ui.exit(1)
     else:
         bibentry_raw = content.get_content(bibfile, ui=ui)
-        bibentry = rp.databroker.verify(bibentry_raw)
+        bibentry = decoder.decode_bibdata(bibentry_raw)
         if bibentry is None:
             ui.error('invalid bibfile {}.'.format(bibfile))
 
@@ -136,25 +138,20 @@ def command(conf, args):
                     '{}, using {} instead.').format(bib_docfile, docfile))
 
     # create the paper
-    copy = args.copy
-    if copy is None:
-        copy = conf['main']['doc_add'] in ('copy', 'move')
-    move = args.move
-    if move is None:
-        move = conf['main']['doc_add'] == 'move'
+    doc_add = args.doc_copy
+    if doc_add is None:
+        doc_add = conf['main']['doc_add']
 
     rp.push_paper(p)
     ui.message('added to pubs:\n{}'.format(pretty.paper_oneliner(p)))
     if docfile is not None:
-        rp.push_doc(p.citekey, docfile, copy=copy or args.move)
-        if copy:
-            if move:
-                content.remove_file(docfile)
+        rp.push_doc(p.citekey, docfile, copy=(doc_add in ('copy', 'move')))
+        if doc_add == 'move' and content.content_type(docfile) != 'url':
+            content.remove_file(docfile)
 
-        if copy:
-            if move:
-                ui.message('{} was moved to the pubs repository.'.format(docfile))
-            else:
+        if doc_add == 'move':
+            ui.message('{} was moved to the pubs repository.'.format(docfile))
+        elif doc_add == 'copy':
                 ui.message('{} was copied to the pubs repository.'.format(docfile))
 
     rp.close()